requirements.txt
django
djangorestframework
requests-html